requirements.txt
anyio==3.7.0
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.0
beartype==0.14.1
build==0.10.0
CacheControl==0.12.14
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
cryptography==37.0.4
dagger-io==0.9.3
distlib==0.3.6
dnspython==2.3.0
dulwich==0.21.5
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.80.0
filelock==3.12.1
gql==3.4.1
graphql-core==3.2.3
h11==0.13.0
html5lib==1.1
httpcore==0.17.2
httpx==0.24.1
idna==3.3
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
motor==3.0.0
msgpack==1.0.5
multidict==6.0.4
odmantic==0.6.0
packaging==23.1
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.5.3
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
Pygments==2.15.1
pymongo==4.3.3
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-jose==3.3.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.18.3
virtualenv==20.23.0
webencodings==0.5.1
xattr==0.10.1
yarl==1.9.2
zipp==3.15.0
